refactor: 구현의도에 맞는 DAO 명 수정
