+++ src/main/java/chess/Application.java
@@ -1,8 +1,8 @@
 package chess;
 
 import chess.controller.ChessController;
+import chess.dao.BoardDaoImpl;
 import chess.dao.BoardDao;
-import chess.dao.BoardRepository;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -10,8 +10,8 @@ public class Application {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        BoardRepository boardRepository = new BoardDao();
-        ChessController chessController = new ChessController(inputView, outputView, boardRepository);
+        BoardDao boardDao = new BoardDaoImpl();
+        ChessController chessController = new ChessController(inputView, outputView, boardDao);
         chessController.run();
     }
 }
