--- src/main/java/chess/dao/BoardRepository.java
@@ -1,13 +1,0 @@
-package chess.dao;
-
-import chess.dto.ChessBoardDto;
-
-public interface BoardRepository {
-    void addAll(final ChessBoardDto chessBoardDto);
-
-    ChessBoardDto findAll();
-
-    int count();
-
-    void deleteAll();
-}
